destination file field removed.
From now on, the destination will be asked to the user once he validates
the changes.
The application also resists when the user tries to validate without
filling one of the required fields. (there is a small problem with the
popup window though. I suspect JavaFX to actually be the problem)


Signed-off-by: Polytrope <alexandre78ferrera@gmail.com>